Refactor url proxy helper method out for use by all platforms.
