+++ browser/net/BUILD.gn
@@ -38,6 +38,7 @@ source_set("browser_tests") {
     "//brave/components/brave_shields/core/common",
     "//brave/components/query_filter",
     "//brave/components/tor/buildflags",
+    "//brave/net/static_redirect_helper",
     "//chrome/browser",
     "//chrome/browser:browser_process",
     "//chrome/browser/content_settings:content_settings_factory",
@@ -92,6 +93,7 @@ source_set("unit_tests") {
     "//brave/components/brave_shields/content/browser",
     "//brave/components/brave_shields/content/test:test_support",
     "//brave/components/l10n/common:test_support",
+    "//brave/net/static_redirect_helper",
     "//chrome/browser/prefs",
     "//chrome/test:test_support",
     "//content/test:test_support",
