+++ browser/net/brave_static_redirect_network_delegate_helper.h
@@ -7,15 +7,12 @@
 #define BRAVE_BROWSER_NET_BRAVE_STATIC_REDIRECT_NETWORK_DELEGATE_HELPER_H_
 
 #include <memory>
-#include <string>
 
 #include "brave/browser/net/url_context.h"
 
 struct BraveRequestInfo;
 
 namespace brave {
-
-inline constexpr char kSafeBrowsingTestingEndpoint[] = "test.safebrowsing.com";
 
 int OnBeforeURLRequest_StaticRedirectWork(
     const ResponseCallback& next_callback,
@@ -24,8 +21,6 @@ int OnBeforeURLRequest_StaticRedirectWork(
 int OnBeforeURLRequest_StaticRedirectWorkForGURL(
     const GURL& request_url,
     GURL* new_url);
-
-void SetSafeBrowsingEndpointForTesting(bool testing);
 
 }  // namespace brave
