+++ browser/net/brave_static_redirect_network_delegate_helper_unittest.cc
@@ -11,6 +11,7 @@
 #include "base/strings/string_util.h"
 #include "brave/browser/net/brave_geolocation_buildflags.h"
 #include "brave/browser/net/url_context.h"
+#include "brave/net/static_redirect_helper/static_redirect_helper.h"
 #include "components/component_updater/component_updater_url_constants.h"
 #include "net/base/net_errors.h"
 #include "testing/gtest/include/gtest/gtest.h"
