+++ net/static_redirect_helper/DEPS
@@ -1,0 +1,6 @@
+include_rules = [
+  "+brave/browser/net",
+  "+brave/browser/safebrowsing",
+  "+brave/components/constants",
+  "+extensions/common",
+]
